# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PERSONAL-3 Alpaca Trading System — Operator API
  version: 1.1.0
  license:
    name: Proprietary — PERSONAL-3 operator only
    identifier: LicenseRef-PERSONAL-3-Proprietary
  description: |
    Ганц operator-т зориулсан арилжааны системийн REST API.

    **Бүх хариунд хүчинтэй гурван үл хөдлөх дүрэм:**
    1. Мөнгөн утга бүр **тэмдэгт мөр** хэлбэрийн fixed-point decimal
       (`"10234.56"`). Float ХЭЗЭЭ Ч биш (AC-25).
    2. Timestamp бүр UTC, RFC 3339, `Z` төгсгөлтэй (AC-26).
    3. Өгөгдөлтэй хариу бүр `source` талбартай ба нэг хариу дотор
       ХОЛИМОГ `source` байхыг хориглоно (AC-20, AC-21).

    **Аюулгүйн гэрээ:** энэ API-д Alpaca руу order илгээх ЦОРЫН ГАНЦ зам нь
    `POST /orders/manual` ба approval queue-ийн `approve` хоёр. Хоёул дотооддоо
    Risk Agent → Execution модулиар дамжина. LLM нь энэ API-д огт хүрэхгүй —
    LLM зөвхөн Tool Contract (энэ файлын 3-р баримт) руу ханддаг.
servers:
  - url: http://localhost:8000/api/v1
    description: DEV (paper)
  - url: https://{host}/api/v1
    description: UAT / Prod
    variables:
      host:
        default: trader.internal
...
security:
  - operatorSession: []
paths:
...
  /kill-switch:
    post:
      operationId: postKillSwitch
      tags: [system]
      summary: ЗОГСОО — шинэ order илгээхийг ТЭР ДОР НЬ зогсоох (FR-7, AC-14)
      description: |
        Төлөв → `halted`, 1 секунд дотор. Байгаа позицыг АВТОМАТААР ХААХГҮЙ
        (спек A-2). Идемпотент: аль хэдийн `halted` бол 200 буцаана.
        Энэ дуудалт баталгаажуулалт ШААРДАХГҮЙ — яаралтай хаалтын замд
        саад тавихгүй.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Зогссон
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStateEnvelope"
              example:
                source: alpaca_paper
                as_of: "2026-09-16T14:31:00Z"
                stale: false
                system_state: halted
                state: halted
                reason: "Operator kill switch"
                changed_at: "2026-09-16T14:31:00Z"
                changed_by: operator
                wind_down_deadline: null
                seconds_remaining: null
                breaker_metrics: []
  /system/wind-down:
    post:
      operationId: postSystemWindDown
      tags: [system]
      summary: УНТРААХ БЭЛТГЭЛ — эхлээд хаах цонх, дараа нь зогсох (FR-13, AC-34)
      description: |
        Төлөв → `winding_down`, 1 секунд дотор, `system` WS сувгаар түгээнэ.
        `wind_down_deadline = now + WIND_DOWN_GRACE` (хуанлийн цагаар).
        Grace дуусахад автоматаар `halted` (AC-36).

        Энэ үед: эрсдэл нэмэгдүүлэх бүх order татгалзана; хаах / багасгах
        order дамжина. Позиц АВТОМАТААР хаагдахгүй (спек A-7) — agent ба
        operator-т хаах БОЛОМЖ өгч байгаа, албадан зарахгүй.
      responses:
        "200":
          description: Wind-down эхлэв
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStateEnvelope"
              example:
                source: alpaca_paper
                as_of: "2026-09-16T14:30:00Z"
                stale: false
                system_state: winding_down
                state: winding_down
                reason: "Operator: машинаа унтраах гэж байна"
                changed_at: "2026-09-16T14:30:00Z"
                changed_by: operator
                wind_down_deadline: "2026-09-16T14:45:00Z"
                seconds_remaining: 900
                breaker_metrics: []
        "409":
          description: Аль хэдийн `halted` — wind-down утгагүй
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /system/activate:
    post:
      operationId: postSystemActivate
      tags: [system]
      summary: ИДЭВХЖҮҮЛ — `ACTIVE` руу буцах ЦОРЫН ГАНЦ зам (AC-37)
      description: |
        Хоёр шаттай баталгаажуулалт ЗААВАЛ. Түүнчлэн backend нь circuit
        breaker-ийн бүх метрикийг ДАХИН ҮНЭЛНЭ — аль нэг нь босгоосоо дээр
        хэвээр бол 409, аль метрик болохыг нэрлэж буцаана. Энэ нь
        «цэвэрлэсэн» гэсэн туг БИШ, БОДИТ хэмжилт — тугийг гараар асаах
        замаар тойрох боломжгүй.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confirmation_token:
                  type: string
                note:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Идэвхжив
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemStateEnvelope"
        "409":
          description: |
            `confirmation_required` — token дутуу; эсвэл
            `breaker_still_tripped` — метрик нэрлэгдэнэ; эсвэл
            `already_active`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    operatorSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Ганц operator-ийн session cookie (HttpOnly, Secure, SameSite=Strict).
        Олон хэрэглэгчийн role model v1-д БАЙХГҮЙ (спек A-6).
...
  schemas:
...
    Utc:
      type: string
      format: date-time
      pattern: 'Z$'
      description: RFC 3339, ЗААВАЛ UTC (`Z`). Локал бүс хориотой (AC-26).
    Source:
      type: string
      enum: [alpaca_live, alpaca_paper, backtest]
      description: |
        Энэ хариун дахь БҮХ өгөгдлийн гарал. Нэг хариу дотор холимог утга
        БАЙХГҮЙ — холих шаардлагатай бол тусдаа хүсэлт (AC-20, AC-21).
    SystemState:
      type: string
      enum: [active, winding_down, halted]
...
    RiskDecision:
      type: string
      enum: [APPROVE, REJECT, ESCALATE_TO_HUMAN]
...
    Envelope:
      type: object
      required: [source, as_of, stale, system_state]
      properties:
        source:
          $ref: "#/components/schemas/Source"
        as_of:
          $ref: "#/components/schemas/Utc"
        stale:
          type: boolean
          default: false
          description: |
            Өгөгдөл `STALE_AFTER_SECONDS`-оос хуучин. Талбарыг ОРХИХГҮЙ —
            ил `false` бичнэ (хавсралт 10 §2.4: «байхгүй» нь таамаглах цонх).
        system_state:
          $ref: "#/components/schemas/SystemState"
...
    RiskEvaluation:
      type: object
      required: [decision, checks, evaluated_at]
      properties:
        decision:
          $ref: "#/components/schemas/RiskDecision"
        reason:
          type: [string, "null"]
        evaluated_at:
          $ref: "#/components/schemas/Utc"
        checks:
          type: array
          description: >-
            Шалгалт бүрийн үр дүн — юу унасныг, ямар хязгаартай харьцуулснаар
            нь ил харуулна.
          items:
            type: object
            required: [rule, passed]
            properties:
              rule:
                type: string
                enum:
                  - system_state
                  - restricted_symbol
                  - position_pct
                  - total_exposure_pct
                  - daily_loss_limit
                  - pdt_day_trades
                  - order_notional
                  - price_sanity
                  - wind_down_direction
              passed:
                type: boolean
              limit_name:
                type: string
                description: "config-ийн нэр, жишээ: `MAX_POSITION_PCT`"
              limit_value:
                type: string
              actual_value:
                type: string
              detail:
                type: string
...
    SystemStateEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          state: active
          reason: null
          changed_at: "2026-09-16T09:00:00Z"
          changed_by: operator
          wind_down_deadline: null
          seconds_remaining: null
          breaker_metrics:
            - metric: daily_loss
              value: "-312.44"
              limit_name: MAX_DAILY_LOSS
              limit_value: "-2000.00"
              tripped: false
            - metric: api_error_rate
              value: "0.004"
              limit_name: MAX_API_ERROR_RATE
              limit_value: "0.05"
              tripped: false
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [state, changed_at, changed_by]
          properties:
            state:
              $ref: "#/components/schemas/SystemState"
            reason:
              type: [string, "null"]
            changed_at:
              $ref: "#/components/schemas/Utc"
            changed_by:
              type: string
              enum: [operator, circuit_breaker, scheduler]
            wind_down_deadline:
              type: [string, "null"]
              format: date-time
              description: Зөвхөн `winding_down` үед. Дуусахад `halted`.
            seconds_remaining:
              type: [integer, "null"]
            breaker_metrics:
              type: array
              description: |
                Circuit breaker-ийн метрик бүрийн одоогийн байдал. `activate`
                энэ жагсаалтыг ДАХИН хэмжиж шийднэ — хадгалагдсан туг биш.

                Хэмжигдээгүй метрик нь `value: "unmeasured"` гэж ИЛ гарна
                (broker хүрэхгүй, эсвэл цонхонд нэг ч үйл явдал байхгүй).
                Хэмжигдээгүйг `0` гэж харуулахгүй.
              items:
                type: object
                required: [metric, value, limit_name, limit_value, tripped]
                properties:
                  metric:
                    type: string
                    enum: [daily_loss, api_error_rate, order_reject_rate, ws_disconnects]
                  value:
                    type: string
                    description: |
                      Тоон утга эсвэл `unmeasured`. `unmeasured` үед `tripped`
                      нь ҮРГЭЛЖ `false` — хэмжээгүй зүйл халт үүсгэхгүй.
                  limit_name:
                    type: string
                  limit_value:
                    type: string
                  tripped:
                    type: boolean
...
    Problem:
      description: RFC 9457 Problem Details.
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - confirmation_required
            - system_halted
            - winding_down_increase_blocked
            - risk_rejected
            - grounding_failed
            - idempotency_conflict
            - breaker_still_tripped
            - already_active
            - approval_expired
            - version_conflict
            - broker_unavailable
            - broker_rejected
            - provider_unavailable
            - not_found
            - invalid_request
        detail:
          type: string
        broker_code:
          type: string
          description: >-
            Зөвхөн `broker_rejected` үед. Alpaca-ийн алдааны код
            (ж: `42210000` — wash trade). Broker ХАРИУЛСАН гэдгийн нотолгоо.
        errors:
          type: array
          description: Зөвхөн `invalid_request` үед. Унасан талбар тус бүр.
          items:
            type: object
            required: [loc, msg]
            properties:
              loc:
                type: string
              msg:
                type: string
        risk:
          $ref: "#/components/schemas/RiskEvaluation"
        confirmation:
          type: object
          description: >-
            Зөвхөн `confirmation_required` үед. Хоёр шаттай баталгаажуулалтын
            сорилт.
          required: [token, prompt, expires_at]
          properties:
            token:
              type: string
            prompt:
              type: string
              description: |
                Operator-т харуулах тодорхой текст. Үйлдэл тус бүрд ӨӨР —
                хоёр товчийг андуурахаас сэргийлнэ (AC-38).
            expires_at:
              $ref: "#/components/schemas/Utc"
        breaker_metrics:
          type: array
          items:
            type: object
            additionalProperties: true